# rkv/src/server/routes/ui/openapi.yaml
openapi: 3.1.0
info:
  title: rKV
  description: Revision-aware key-value store with LSM-tree storage engine.
  version: 0.1.0
  license:
    name: MIT
servers:
  - url: /
paths:
...
  /api/{ns}/batch:
    post:
      tags: [Batch]
      summary: Atomic batch write
      operationId: writeBatch
      parameters:
        - name: ns
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ops]
              properties:
                ops:
                  type: array
                  items:
                    oneOf:
                      - type: object
                        required: [op, key, value]
                        properties:
                          op:
                            type: string
                            enum: [put]
                          key:
                            type: string
                          value: {}
                          ttl:
                            type: integer
                            description: TTL in seconds
                          dedup:
                            type: boolean
                            description: Per-op dedup override
                      - type: object
                        required: [op, key]
                        properties:
                          op:
                            type: string
                            enum: [delete]
                          key:
                            type: string
            example:
              ops:
                - op: put
                  key: k1
                  value: v1
                - op: put
                  key: k2
                  value: v2
                  ttl: 3600
                  dedup: true
                - op: delete
                  key: k3
      responses:
        "200":
          description: Revision IDs for each operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        key:
                          type: string
                        revision:
                          type: string
...